+++ MasterThesis/RePair_v2/GammaCode.h
@@ -5,29 +5,29 @@ class GammaCode
 public:
 	void GammaCode::decodeGammaString(std::string &prefix, 
 		std::string &gamma,
-		std::vector<unsigned int> &output, 
-		unsigned int count);
-	unsigned int GammaCode::binaryToInt(std::string binary);
-	unsigned int GammaCode::readGammaCodeHeader(std::string& gamma, int& i);
-	unsigned int GammaCode::gammaToInt(std::string binary);
-	std::string getBinaryCode(unsigned int input);
-	std::string getGammaCode(unsigned int input);
+		std::vector<long> &output, 
+		long count);
+	long GammaCode::binaryToInt(std::string binary);
+	long GammaCode::readGammaCodeHeader(std::string& gamma, int& i);
+	long GammaCode::gammaToInt(std::string binary);
+	std::string getBinaryCode(long input);
+	std::string getGammaCode(long input);
 	void makeFinalString(std::vector<std::vector<CompactPair>>& pairVector,
-		std::unordered_set<unsigned int>& terminals,
+		std::unordered_set<long>& terminals,
 		std::string& finalString,
 		std::vector<std::vector<CompactPair>> generationVectors);
 	void encode(std::vector<std::vector<CompactPair>>& pairVector,
-		std::unordered_set<unsigned int>& terminals,
+		std::unordered_set<long>& terminals,
 		std::string& terminalsGamma,
 		std::vector<std::string>& leftElementsGammas,
 		std::vector<std::string>& rightElementsBinaries,
 		std::vector<std::vector<CompactPair>>& generationVectors);
 	void decode(std::vector<std::vector<CompactPair>>& pairVector,
-		std::unordered_set<unsigned int>& terminals,
+		std::unordered_set<long>& terminals,
 		std::string& inputString);
-	void GammaCode::readNextNumbers(int n, std::vector<unsigned int> &values, std::ifstream &bitstream, std::string &prefix);
-	void GammaCode::readNextBinaries(int binarySize, int count, std::vector<unsigned int> &values, std::ifstream &bitstream, std::string &prefix);
+	void GammaCode::readNextNumbers(int n, std::vector<long> &values, std::ifstream &bitstream, std::string &prefix);
+	void GammaCode::readNextBinaries(int binarySize, int count, std::vector<long> &values, std::ifstream &bitstream, std::string &prefix);
 	void GammaCode::decodeDictionaryFile(std::vector<CompactPair>& pairs,
-		std::unordered_set<unsigned int>& terminals,
+		std::unordered_set<long>& terminals,
 		std::ifstream &bitstream);
 };
